requirements.txt
google-cloud-storage==2.10.0
ipykernel==6.24.0
opencv-python==4.8.0.74
matplotlib==3.7.2